requirements.txt
asgiref
Django==3.2
pytz==2022.1
sqlparse==0.4.2
djangorestframework==3.13.1